+++ airflow/contrib/sensors/sagemaker_endpoint_sensor.py
@@ -16,46 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.sagemaker_endpoint`."""
 
-from airflow.contrib.hooks.sagemaker_hook import SageMakerHook
-from airflow.providers.amazon.aws.sensors.sagemaker_base import SageMakerBaseSensor
-from airflow.utils.decorators import apply_defaults
+import warnings
 
+# pylint: disable=unused-import
+from airflow.providers.amazon.aws.sensors.sagemaker_endpoint import SageMakerEndpointSensor  # noqa
 
-class SageMakerEndpointSensor(SageMakerBaseSensor):
-    """
-    Asks for the state of the endpoint state until it reaches a terminal state.
-    If it fails the sensor errors, the task fails.
-
-    :param job_name: job_name of the endpoint instance to check the state of
-    :type job_name: str
-    """
-
-    template_fields = ['endpoint_name']
-    template_ext = ()
-
-    @apply_defaults
-    def __init__(self,
-                 endpoint_name,
-                 *args,
-                 **kwargs):
-        super().__init__(*args, **kwargs)
-        self.endpoint_name = endpoint_name
-
-    def non_terminal_states(self):
-        return SageMakerHook.endpoint_non_terminal_states
-
-    def failed_states(self):
-        return SageMakerHook.failed_states
-
-    def get_sagemaker_response(self):
-        sagemaker = SageMakerHook(aws_conn_id=self.aws_conn_id)
-
-        self.log.info('Poking Sagemaker Endpoint %s', self.endpoint_name)
-        return sagemaker.describe_endpoint(self.endpoint_name)
-
-    def get_failed_reason_from_response(self, response):
-        return response['FailureReason']
-
-    def state_from_response(self, response):
-        return response['EndpointStatus']
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.sagemaker_endpoint`.",
+    DeprecationWarning, stacklevel=2
+)
