+++ scripts/ci/pylint_todo.txt
@@ -93,7 +93,7 @@
 ./airflow/contrib/sensors/redis_key_sensor.py
 ./airflow/contrib/sensors/redis_pub_sub_sensor.py
 ./airflow/providers/amazon/aws/sensors/sagemaker_base.py
-./airflow/contrib/sensors/sagemaker_endpoint_sensor.py
+./airflow/providers/amazon/aws/sensors/sagemaker_endpoint.py
 ./airflow/contrib/sensors/sagemaker_training_sensor.py
 ./airflow/contrib/sensors/sagemaker_transform_sensor.py
 ./airflow/contrib/sensors/sagemaker_tuning_sensor.py
